[refactor] App: migrate old app's domain and data layer to new app
