+++ app/src/main/java/com/rwmobi/dazncodechallenge/domain/exceptions/RunCatchingException.kt
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c) 2024. Ryan Wong
+ * https://github.com/ryanw-mobile
+ * Sponsored by RW MobiMedia UK Limited
+ *
+ */
+
+package com.rwmobi.dazncodechallenge.domain.exceptions
+
+/***
+ * Workaround to rethrow CancellationException
+ * Reference: https://github.com/Kotlin/kotlinx.coroutines/issues/1814
+ */
+
+inline fun <reified T : Throwable, R> Result<R>.except(): Result<R> =
+    onFailure { if (it is T) throw it }
