--- old version/app/src/main/java/uk/ryanwong/dazn/codechallenge/data/source/local/DaznApiDatabase.kt
@@ -1,17 +1,0 @@
-package uk.ryanwong.dazn.codechallenge.data.source.local
-
-import androidx.room.Database
-import androidx.room.RoomDatabase
-import androidx.room.TypeConverters
-import uk.ryanwong.dazn.codechallenge.data.source.local.daos.EventsDao
-import uk.ryanwong.dazn.codechallenge.data.source.local.daos.ScheduleDao
-import uk.ryanwong.dazn.codechallenge.data.source.local.entities.EventDbEntity
-import uk.ryanwong.dazn.codechallenge.data.source.local.entities.ScheduleDbEntity
-
-@Database(entities = [EventDbEntity::class, ScheduleDbEntity::class], version = 5, exportSchema = false)
-@TypeConverters(RoomDbConverters::class)
-abstract class DaznApiDatabase : RoomDatabase() {
-    // DAOs
-    abstract val eventsDao: EventsDao
-    abstract val scheduleDao: ScheduleDao
-}
