--- old version/app/src/main/java/uk/ryanwong/dazn/codechallenge/data/source/remote/DaznApiService.kt
@@ -1,18 +1,0 @@
-/*
- * Copyright (c) 2021. Ryan Wong (hello@ryanwong.co.uk)
- *
- */
-
-package uk.ryanwong.dazn.codechallenge.data.source.remote
-
-import retrofit2.http.GET
-import uk.ryanwong.dazn.codechallenge.data.source.remote.entities.EventNetworkEntity
-import uk.ryanwong.dazn.codechallenge.data.source.remote.entities.ScheduleNetworkEntity
-
-sealed interface DaznApiService {
-    @GET("getEvents")
-    suspend fun getEvents(): List<EventNetworkEntity>
-
-    @GET("getSchedule")
-    suspend fun getSchedule(): List<ScheduleNetworkEntity>
-}
