--- old version/app/src/main/java/uk/ryanwong/dazn/codechallenge/domain/models/Schedule.kt
@@ -1,11 +1,0 @@
-package uk.ryanwong.dazn.codechallenge.domain.models
-
-import java.util.Date
-
-data class Schedule(
-    val scheduleId: Int,
-    val title: String,
-    val subtitle: String,
-    val date: Date,
-    val imageUrl: String,
-)
