--- old version/app/src/main/java/uk/ryanwong/dazn/codechallenge/util/extensions/ParseTimeStamp.kt
@@ -1,15 +1,0 @@
-/*
- * Copyright (c) 2021. Ryan Wong (hello@ryanwong.co.uk)
- *
- */
-
-package uk.ryanwong.dazn.codechallenge.util.extensions
-
-import java.text.SimpleDateFormat
-import java.util.Date
-import java.util.Locale
-
-fun String.parseTimeStamp(dateFormat: String = "yyyy-MM-dd'T'HH:mm:ss"): Date {
-    val simpleDateFormat = SimpleDateFormat(dateFormat, Locale.getDefault())
-    return simpleDateFormat.parse(this)!!
-}
