+++ old version/app/src/main/res/layout/fragment_events.xml
@@ -14,7 +14,7 @@
 
         <variable
             name="viewModel"
-            type="uk.ryanwong.dazn.codechallenge.ui.events.EventsViewModel" />
+            type="com.rwmobi.dazncodechallenge.ui.viewmodel.EventsViewModel" />
     </data>
 
     <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
