+++ old version/app/src/main/res/layout/listitem_events.xml
@@ -13,7 +13,7 @@
 
         <variable
             name="event"
-            type="uk.ryanwong.dazn.codechallenge.domain.models.Event" />
+            type="com.rwmobi.dazncodechallenge.domain.model.Event" />
 
         <variable
             name="clickListener"
